Add custom LTI field ext_context_locale
LTI has a launch_presentation_locale field which is specific to the user launching the tool.
The ext_context_locale custom field is useful when the LTI instance is shared between course users. For example, a chat room or web conference.
A tool consumer is allowed to send custom LTI fields prefixed with ext_. Moodle sends the custom ext_user_username and ext_lms fields.